Add array JSON I/O
