+++ app/demo-rasterizer/ImageIO.cc
@@ -7,6 +7,7 @@
 //---------------------------------------------------------------------------//
 #include "ImageIO.hh"
 
+#include "base/Array.json.hh"
 #include "ImageStore.hh"
 
 namespace demo_rasterizer
