+++ app/demo-rasterizer/ImageIO.hh
@@ -32,30 +32,3 @@ void from_json(const nlohmann::json& j, ImageRunArgs& value);
 void to_json(nlohmann::json& j, const ImageStore& value);
 //---------------------------------------------------------------------------//
 } // namespace demo_rasterizer
-
-namespace celeritas
-{
-//---------------------------------------------------------------------------//
-template<class T, size_type N>
-void from_json(const nlohmann::json& j, Array<T, N>& value)
-{
-    CELER_EXPECT(j.size() == N);
-    for (size_type i = 0; i != N; ++i)
-    {
-        value[i] = j[i].get<T>();
-    }
-}
-
-//---------------------------------------------------------------------------//
-template<class T, size_type N>
-void to_json(nlohmann::json& j, const Array<T, N>& value)
-{
-    j = nlohmann::json::array();
-    for (size_type i = 0; i != N; ++i)
-    {
-        j.push_back(value[i]);
-    }
-}
-
-//---------------------------------------------------------------------------//
-} // namespace celeritas
